+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - feat: add command `update-single-source` to update a single source assuming all the episodes are new
 - feat: add command `print` to print the sources definitions and source names
 
+### Fixed
+
+- fix: do not stop execution when a provider fails to fetch its episodes
+
 ## [1.0.0] - 2023-03-05
 
 ### Added